XMLValidationTestCase: criterion correctly defines implict validity.
A filename: "illegal_syntax.xml" was giving an expected result of VALID 
for criterion='syntax' when it should be ILLEGAL. The substring 'syntax' 
is a signal that the file would yield a syntax-related exception and not 
that it would pass a syntax validator. Quite the opposite in fact. This 
'Truthy' notion was being confused.